internal/api: implement basic client
Before implementing a full terraform provider, we need to build out a
client library for interacting with the sitehost api. This change adds
the following basic interfaces to get started with a simple terraform
provider: /api/get_info.json, /job/get.json, /server/delete.json,
/server/get_server.json, /server/list_servers.json, and
/server/provision.json.

Query and form parameter ordering is important, so that is hardcoded for
each api call.

In addition to mocked unit tests, one integration test that interacts
with production services was also added. It can be skipped by passing
the -test.short flag to the toolchain.